receive file information in the content process
we're disabling two security features here that started being on-by-defualt in electron v12. They assume that most implementers will be loading some remote webpage in the content process and want to prevent that webpage from having arbitrary local access. We're building our own page here and not loading remote or 3rd party content (aside from the PGN file off disk, which will be parsed). Therefor this doesn't really help us mitigate any risk and instead we would have to complicate our application, somewhat. If we did need to do it, though, here's how we would:

create a preload script, which can access `ipcRenderer` and create an object that passes to the content script namespace. Create our API in that preload script and pass an object in that is capable of receiving and sending over ipc.

It's a bit of indirection, but is maybe more future proof. If we start loading any content other that pgn data we should consider this.