stubbed out UI
this commit stubbs out the non-board UI that we've been missing. This includes the game metadata, displayed in the header or integrated into the board, as well as the basic layout and position of the move list.

the move list does not dynamically resize. In the future I think we will want to put some min-max bounds on the game board and think about how the UI list should flow from there. I do most of my testing in a vertical-cell-phone-like-orientation but should consider wider formats