introducting htm
preact has a precompile step but that feels like a lot to set up right now, so instead we're going to use the htm library to work with preact in a no build, no jsx manner.

this will let us use Tagged Templates for rendering the UI

eventually we may want to add babel and weplace all this with templates but let's defer that as far as we comfortably can.